pipeline/pyproject.toml
[project]
name = "frostbyte-pipeline"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.10",
    "python-multipart>=0.0.9",
    "boto3>=1.35",
    "asyncpg>=0.29",
    "qdrant-client>=1.13",
    "httpx>=0.27",
    "redis>=5.0",
    "python-magic>=0.4.27",
    "python-jose[cryptography]>=3.3",
    "clamd>=1.0.2",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["pipeline*"]
